+++ CHANGELOG.md
@@ -16,6 +16,10 @@ FEATURES:
 * VMs can be searched by ID or by Name, Organization, VDC and vApp
 * Add support for customization scripts
 * Add commands to manage VM's status (start/stop/delete/reset/suspend/reboot)
+* Add command to show details about a given network
+
+VARIOUS:
+* Update dependency vcloud-rest v. 0.4.0
 
 2012-12-28 (0.2.3)
 --